Formulario CLIENTE OK Formulado Pedido OK Formulado Material OK

+++ src/modelos/Material.java
@@ -1,15 +1,17 @@
 package modelos;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author babar
  */
 public class Material {
     private String tipo;
-    private double valor;  
+    private BigDecimal valor;  
    
 
-    public Material(String tipo, double valor) {
+    public Material(String tipo, BigDecimal valor) {
         this.tipo = tipo;
         this.valor = valor;
     }
@@ -22,11 +24,11 @@ public class Material {
         this.tipo = tipo;
     }
 
-    public double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
